Add quick eval methods to AST, no Promises.
Add quick eval (qeval) computed properties to the AST for Expression
nodes. This methods behave identically to the standard evalfn properties
but do not use Promises. qeval can also be null in the case that it
can't be guaranteed that a promise is not needed.

At this time the only case when a Promise might be required is if there
is a function call.

Also update expression tests to use both evaluation strategies