Add syntax highlight to docs with setanta-hl
Use setanta-hl binary to convert markdown setanta code blocks to syntax
highlighted blocks.

This is a work around the rouge highlighting system, as we would have to
get a PR submitted to rouge and wait for Github pages to update to
latest version to get Setanta highlighting directly. I'd also have to
learn Ruby.

To get around this I wrote a custom syntax preprocessor that would parse
the markdown files and generate highlighted html. The css classes are
derived from the ones that rouge would generate itself.

This is extremely hacky, hopefully it works ok