Update REPL to support multiple lines
If a line entered into the REPL fails to parse, and the failure location
is at the end of the line, then open a new line and allow more to be
typed to finish the line.

This allows you to write multiline functions, classes etc. Don't have to
compress them onto one line anymore