Add rejection pool for outstanding promises
Add a pool of promise rejection functions to the global context that
will be called when the interpreter is stopped.
Needs to be used for promises which are handled outside of the usual
interpreter lifecycle, e.g. setTimeout calls.
When interpreter is stopped, all rejections functions in the pool are
called.