Fix variable binding bug (oops)
bug introduced with lexical scoping introduction in 0.9.0 was caused by
false self-reference detection. If a variable was defined as an object
lookup (e.g. `x := a@b`) but one of the members of the object has the
same name as the variable, this was incorrectly flagged as a self
reference failure (e.g. `x := x@b`).

This broke the main snake script on the try-setanta site

This commit fixes the issue and adds many regression tests